Move ParsedVectorQuery to vector_query_utils; update includes in relevant files
